add AbstractEventStoreTest
see: https://github.com/prooph/event-store/issues/257
